I forgot to save tooltip specification changes to file before last commit, so these are in this commit.
Added header bar. This is a bar above the bible frame which displays the
current chapter and the ones around it. Hovering over a chapter in it will
give the chapter, number of verses, and list of headings in the current chapter, each hyperlinked for easy access.

Removed the html header which used to be there, with its forward and back
buttons.

Added chapter_headings.py in the backend for the retrieval of headings from
the current Bible to support the header bar.



git-svn-id: https://bpbible.googlecode.com/svn/trunk@56 20523bf7-2b47-0410-960a-a116b6c96ec9