Reworked paths handling code to have one SWMgr per path. Made book work with unicode.
Removed book search, which isn't used.

Added lazy topic list, which precomputes number of keys in a dictionary, and
then will get keys when asked for, putting them in a cache. It also puts a
certain number around the key asked for (currently 10, but needs tuning). It
falls back on the old method if it encounters an error.

Changed method of registration of new biblemgr for filterutils so that it
works properly with the new biblemgr.

Changes throughout for handling unicode.



git-svn-id: https://bpbible.googlecode.com/svn/branches/installmgr@77 20523bf7-2b47-0410-960a-a116b6c96ec9