Fixed bug where options which weren't boolean weren't working.
Renamed options menu to Bible Options, as Daniel suggested. I may change this
back if I think it isn't as good.

Changes throughout for unicode.


Changed PathManager to a panel, so that it can be put in the book manager.

In treecombo, changed renamed variable tk to data, as it is not only for
TreeKeys.

Removed checks to see whether we are running a unicode build of wx.
From now on, Unicode builds are mandatory.

Reworked dictionary list and virtuallistbox to support new dictionary topic
list.



git-svn-id: https://bpbible.googlecode.com/svn/branches/installmgr@78 20523bf7-2b47-0410-960a-a116b6c96ec9