Added a smarter template for the bible. This finds any leading whitespace in $text, and puts it at the front of the body item. Any trailing whitespace goes to the back of the body item.
This makes tags appear on the line of the verse if the verse ends with a new
line or new paragraph.

Also, made the bible templates (but not the verse compare ones) so that they
display tags, instead of including tags as before.

Made the verse_per_line linebreak be removed if it is next to whitespace of
the new line variety to stop too many line breaks occurring.



git-svn-id: https://bpbible.googlecode.com/svn/trunk@220 20523bf7-2b47-0410-960a-a116b6c96ec9